Fixed velocity calculations for player movement and also fixed buggy collisons at high-speeds
Now the player's velocity correctly calculates also while on the ground
and all the available modes have the velocity correctly updated.
Refactored and fixed bugs in PlayerStates. Also fixed collisions by
refactoring the isCollidingAt() function in the Player class, the only
downside is that it's a bit longer to calculate and the FPS tanked like
.2%.
---BUGS FIXED---
YEAHHH!!! :)