Added a basic UI system w/RmlUI (a HTML and CSS based UI library)
(No Chromium was used in the makign of the system [THANK GOD!!]) ._.
It's preatty nice but still very unperformant (for ex. recalculating the
layour every frame) the UI is to be designed.
Also custom text rendering with  FreeType's text rendering API.
The UI design rn was taken from RmlUI's sample and is to be cahnged to
adapt to this game.
Overall the system is still very unstable and in very early
developement. Optimizations and system-breaking changes will be
implemented in the future but for now (adn today) it's enough.
YEAH!!!! NEW ACCOMPLISHMENT!!! YAY ;)